Safari-609.1.11.1/LayoutTests/media/video-muted-holds-sleep-assertion-expected.txt


RUN(video.src = findMediaFile("video", "content/test"))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (internals.elementIsBlockingDisplaySleep(video) == 'false') OK
END OF TEST
